rename `ghost-url` service to `ghost` service
