+++ src/controllers/sso.ts
@@ -2,7 +2,7 @@ import {Buffer} from 'node:buffer';
 import {createHmac, timingSafeEqual} from 'node:crypto';
 import {Request, Response} from 'express';
 import fetch from 'node-fetch';
-import {getGhostUrl} from '../services/ghost-url.js';
+import {getGhostUrl} from '../services/ghost.js';
 import {discourseSecret} from '../services/config.js';
 import {GhostMember} from '../types/ghost.js';
